Merging installed SDKs with installable SDKs.
Previously I assumed that the list of SDKs from Appcelerator was
complete; ie, that it listed every build ever. It turns out, however,
that they remove old SDKs; so old SDKs wouldn't show up when doing a tsm
ls all.  This fixes the problem by merging the list of installable sdks
with the list of installed sdks.